refact(135898): refatorado teste repeticao 2 sobremesa c/ usuario emef
