+++ src/mocks/medicaoInicial/PeriodoLancamentoMedicaoInicial/CEUGESTAO/permissoesLancamentosEspeciais.jsx
@@ -16,16 +16,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
         name: "repeticao_2_refeicao",
         uuid: null,
       },
-      {
-        nome: "2ª Sobremesa 1ª oferta",
-        name: "2_sobremesa_1_oferta",
-        uuid: null,
-      },
-      {
-        nome: "Repetição 2ª Sobremesa",
-        name: "repeticao_2_sobremesa",
-        uuid: null,
-      },
     ],
     permissoes_por_dia: [
       {
@@ -35,8 +25,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -47,8 +35,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -59,8 +45,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -71,8 +55,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -83,8 +65,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -95,8 +75,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -107,8 +85,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -119,8 +95,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -131,8 +105,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -143,8 +115,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -155,8 +125,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -167,8 +135,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -179,8 +145,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -191,8 +155,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -203,8 +165,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -215,8 +175,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -227,8 +185,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -239,8 +195,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -251,8 +205,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -263,8 +215,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -275,8 +225,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -287,8 +235,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -299,8 +245,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -311,8 +255,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -323,8 +265,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -335,8 +275,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -347,8 +285,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -359,8 +295,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -371,8 +305,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
@@ -383,8 +315,6 @@ export const mockPermissoesLancamentosEspeciaisCEUGESTAO = {
           "2_refeicao_1_oferta",
           "repeticao_2_refeicao",
           "lanche_extra",
-          "2_sobremesa_1_oferta",
-          "repeticao_2_sobremesa",
         ],
         permissao_id_externo: "30603",
       },
